8_lekcia/requirements.txt
openai>=1.0.0
pyodbc>=4.0.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
